# [perf]fix current stream