fix the performance regression reported from https://github.com/vllm-project/vllm/pull/11744#issuecomment-2578717013

on my local benchmark:

`python benchmarks/benchmark_latency.py --model meta-llama/Meta-Llama-3-70B --load-format dummy --enforce-eager -tp 4`

main branch:

```text
Avg latency: 2.945735554069203 seconds
10% percentile latency: 2.924619035271462 seconds
25% percentile latency: 2.937671729727299 seconds
50% percentile latency: 2.9460502695292234 seconds
75% percentile latency: 2.955668824230088 seconds
90% percentile latency: 2.9639973257959356 seconds
99% percentile latency: 2.979829666109872 seconds
```

this PR:

```text
Avg latency: 2.851606635436959 seconds
10% percentile latency: 2.8231707043829375 seconds
25% percentile latency: 2.834942308269092 seconds
50% percentile latency: 2.85484445450129 seconds
75% percentile latency: 2.8674310567148495 seconds
90% percentile latency: 2.872856835933635 seconds
99% percentile latency: 2.875793117735884 seconds
```

it can have 3% perf diff.

Hopefully this can fix the perf regression observed in the benchmark:

<img width="1247" alt="image" src="https://github.com/user-attachments/assets/2f0698a3-88fe-425a-b461-8458e875bf23" />