@youkaichao 2025-01-09
I find measuring the pure forward time makes more sense, it will not be affected by the scheduling, etc:

`VLLM_LOG_BATCHSIZE_INTERVAL=1 python benchmarks/benchmark_latency.py --model meta-llama/Meta-Llama-3-70B --load-format dummy --enforce-eager -tp 4`

main branch:

```text
Batchsize forward time stats (batchsize, count, median_time(ms)): [(8, 4998, 20.77), (256, 40, 28.99)]
```

this PR:

```text
Batchsize forward time stats (batchsize, count, median_time(ms)): [(8, 5027, 20.45), (256, 40, 28.95)]
```

The forward time for every step (batchsize 8) reduces from 20.77ms to 20.45ms.